This has been a change that I've been meaning to make for TWO YEARS; now its done.
Loading all of zdisp.mat into memory to generate the phantom*.mat files consumes a HUGE amount of RAM for some simulations.  It creates problems when running concurrent jobs on the cluster and it prevents running some jobs at all.  

zdisp.mat files have been replaced by zdisp.dat - a generic binary file that can be read sequentially, one time step at a time.  The new format of this file is as follows:

All data is float32 with the format:
    NUM_NODES
    NUM_DIMS (Node ID, X, Y, Z displacements)
    NUM_TIMESTEPS
    The rest of the data are the concatenation of NUM_NODES x NUM_DIMS x NUM_TIMESTEPS.

convert_zdisp_dat.m can be used to convert pre-existing zdisp.mat files -> zdisp.dat.

All changes with the new zdisp.dat files are contained withing mkphantomfromdyna3.m.

StructPost needs to be changed in the FEM code to create this new zdisp.dat file by default.  This still needs to be done, so in the meantime, the conversion function can be used to create the zdisp.dat file.



git-svn-id: https://hogwarts.egr.duke.edu/svn/ultratrack@23 57c3da9d-1720-0410-8437-ae526f62eb58