Fixed bug with Rx apodization
Field II will error if an apodization curve is all 0 weights.  Pete had
a check for this, but I couldn't follow his logic, and it wasn't
accomodating the first case of Rx apod growth correctly, so I replaced
it with a center element ON correction and a warning statement of it
doing such.